+++ libqpdf/Pl_Discard.cc
@@ -12,7 +12,7 @@ Pl_Discard::~Pl_Discard()
 }
 
 void
-Pl_Discard::write(unsigned char* buf, int len)
+Pl_Discard::write(unsigned char* buf, size_t len)
 {
 }
